abstract a base class from NeuralNetwork and RNN
to make a consistent interface for trainAsync/train
